delete vegeta.go
